single Persistence view
