--- crates/autopilot/src/infra/persistence/api/mod.rs
@@ -1,4 +1,0 @@
-pub mod auction;
-pub mod competition;
-pub mod events;
-pub mod quote;
